:truck: moved buy to V1
