+++ src/mappings/shared/index.ts
@@ -1,5 +1,4 @@
 export * from './burn'
-export * from './buy'
 export * from './change'
 export * from './create'
 export * from './emote'
