+++ src/mappings/v2/index.ts
@@ -1,12 +1,13 @@
 import { Interaction, unwrapRemarkV2 as unwrapRemark } from '@kodadot1/minimark/v2'
 
-import { burn, buy, changeIssuer, emote, list } from '../shared'
+import { burn, changeIssuer, emote, list } from '../shared'
 import { unwrap } from '../utils/extract'
 import logger, { pending } from '../utils/logger'
 import { Context } from '../utils/types'
 import { acceptResource } from './accept'
 import { addResource } from './addResource'
 import { base as createBase } from './base'
+import { buy } from './buy'
 import { createCollection } from './create'
 import { lockCollection } from './lock'
 import { mintItem } from './mint'
